[railway][1/n] constraints on railway controller not moving too much
looks bad if user can move the railway too much around, break the
simulation.
To implement this we first need to have edges and then something like

onDrag(x, y)
compute if this x,y is possible (no distance breaking, no curvature
breaking)
if it's not don't call the update railway method

how to check if x,y is possible? should this be a second test curve?
I like this design. Two curves, one with railway, one without.
Maybe the second curve is a property of the edge? Edge doesn't extend
anymore curve but has